check link (#30)
this PR creates a function checkLink() which acts as a wrapper for error handling and SyntheticLinkResult creation (api parseable object for individual link navigation).

Note that the real action/link-navigation happens in navigate(..). Some of the tests used to be navigate tests that were expanded to be checkLink tests. My logic here was that all the navigate functionality is still being tested and since checkLink is basically just a wrapper on top it does not make sense to have separate unit tests. Rather larger tests (similar to how we test logs).